fix(firehose): refactor trait grid filter
